requirements.txt
google-cloud-bigquery==3.13.0
sentence-transformers>=2.2.2,<3.0.0
transformers>=4.41.0,<4.50.0
pandas==2.1.3
python-dotenv==1.0.0
numpy==1.26.2
pytest==7.4.3
pytest-cov==4.1.0
